Moved color types to their own headers
