+++ src/color/color24bit.h
@@ -1,0 +1,18 @@
+#ifndef PINE_COLOR_COLOR24BIT_H
+#define PINE_COLOR_COLOR24BIT_H
+
+namespace Pine
+{
+    struct Color24Bit;
+
+
+
+    struct Color24Bit
+    {
+        unsigned char red;
+        unsigned char green;
+        unsigned char blue;
+    };
+}
+
+#endif // PINE_COLOR_COLOR24BIT_H
