+++ src/color/color4bit.h
@@ -1,0 +1,26 @@
+#ifndef PINE_COLOR_COLOR4BIT_H
+#define PINE_COLOR_COLOR4BIT_H
+
+namespace Pine
+{
+    enum class Color4Bit {
+        Black          = 30,
+        Red            = 31,
+        Green          = 32,
+        Yellow         = 33,
+        Blue           = 34,
+        Magenta        = 35,
+        Cyan           = 36,
+        White          = 37,
+        Bright_Black   = 90,
+        Bright_Red     = 91,
+        Bright_Green   = 92,
+        Bright_Yellow  = 93,
+        Bright_Blue    = 94,
+        Bright_Magenta = 95,
+        Bright_Cyan    = 96,
+        Bright_White   = 97
+    };
+}
+
+#endif // PINE_COLOR_COLOR4BIT_H
