+++ src/color/color8bit.h
@@ -1,0 +1,9 @@
+#ifndef PINE_COLOR_COLOR8BIT_H
+#define PINE_COLOR_COLOR8BIT_H
+
+namespace Pine
+{
+    typedef unsigned char Color8Bit;
+}
+
+#endif // PINE_COLOR_COLOR8BIT_H
